requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
easyocr
pillow
opencv-python-headless
pandas
